feat: add AppData manifest
Every software center that exists allows the user to look at
screenshots and a long description of the application before it is
installed. For most users it allows them to answer the question “Do I
want to install this application?”. Traditionally in Linux
distributions, we have none of this data for the vast majority of our
desktop user-installable applications.

More info:
https://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html